moved midi synthesizers to audio module
