+++ src/framework/audio/internal/synthesizers/zerberus/internal/channel.h
