+++ src/framework/audio/internal/synthesizers/zerberus/internal/controllers.h
