+++ src/framework/audio/internal/synthesizers/zerberus/internal/instrument.h
