+++ src/framework/audio/internal/synthesizers/zerberus/internal/voice.h
