+++ src/framework/audio/internal/synthesizers/zerberus/internal/zerberus.h
