+++ src/framework/audio/internal/synthesizers/zerberus/internal/zone.h
