+++ src/framework/midi/internal/midiconfiguration.h
@@ -19,37 +19,13 @@
 #ifndef MU_MIDI_MIDICONFIGURATION_H
 #define MU_MIDI_MIDICONFIGURATION_H
 
-#include <map>
-
 #include "../imidiconfiguration.h"
-
-#include "modularity/ioc.h"
-#include "iglobalconfiguration.h"
 
 namespace mu::midi {
 class MidiConfiguration : public IMidiConfiguration
 {
-    INJECT(midi, framework::IGlobalConfiguration, globalConfiguration)
-
 public:
-
-    std::vector<io::path> soundFontPaths() const override;
-
-    const SynthesizerState& defaultSynthesizerState() const;
-    const SynthesizerState& synthesizerState() const override;
-    Ret saveSynthesizerState(const SynthesizerState& state) override;
-    async::Notification synthesizerStateChanged() const override;
-    async::Notification synthesizerStateGroupChanged(const std::string& gname) const override;
-
-private:
-
-    io::path stateFilePath() const;
-    bool readState(const io::path& path, SynthesizerState& state) const;
-    bool writeState(const io::path& path, const SynthesizerState& state);
-
-    mutable SynthesizerState m_state;
-    async::Notification m_synthesizerStateChanged;
-    mutable std::map<std::string, async::Notification> m_synthesizerStateGroupChanged;
+    MidiConfiguration();
 };
 }
