+++ src/framework/midi/midimodule.h
@@ -30,7 +30,6 @@ public:
 
     void registerExports() override;
     void registerUiTypes() override;
-    void onInit(const framework::IApplication::RunMode& mode) override;
 };
 }
